# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API Documentation for fetching and managing news articles.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles/fetch": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch and aggregate articles",
                "description": "Fetches articles from multiple external APIs and stores them in the database.",
                "operationId": "8c98ed0fadc77db9f5a1340dd06059ab",
                "responses": {
                    "200": {
                        "description": "Articles fetched and stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": "success",
                                    "message": "Articles fetched and stored successfully",
                                    "articles_fetched": 150
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": "error",
                                    "message": "Failed to fetch articles"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": "error",
                                    "message": "Invalid parameters provided"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch articles from external APIs",
                "description": "Fetches articles from different sources like NewsAPI, Guardian, and New York Times based on the provided source parameter.",
                "operationId": "89181b74a8aa4d580750ede6aa32f92a",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "The source to fetch articles from (e.g., 'newsapi', 'guardian', 'newyorktimes').",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "newsapi"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Optional search keyword to filter articles.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "technology"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles fetched successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Sample Article Title"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "source": {
                                                "type": "string",
                                                "example": "newsapi"
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-24T10:00:00Z"
                                            },
                                            "url": {
                                                "type": "string",
                                                "example": "https://example.com/article"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid source parameter or missing source.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch articles"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error while fetching articles.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching articles"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}